requirements.txt
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
python-dateutil==2.8.2
PyYAML==6.0.1
pytest==7.4.3
